feat: premium document system, security audit fixes, and deploy-time migrations
Documents and printing
- Vector PDF document system (src/lib/documents): letterhead, Page X of Y,
  Urdu font, QR verification, exact decimal money and marks.
- Rebuilt report cards, vouchers, invoices, receipts, payslips, ID cards,
  certificates, datesheets, admit cards, timetables, appointment letters,
  fee certificates, lesson plans, progress reports, seating door sheets,
  visitor passes, expense vouchers, platform invoices and the board packet.
- Branded Excel / PDF / Print / WhatsApp / CSV exports on the data screens;
  files named after their data.

Fixes found on the way
- Removed fabricated data (seating plans, owner insights, demo fallbacks,
  fake bank details, simulated SMS/WhatsApp claims).
- Tax certificates, hostel allocation, exam seating, JazzCash flow and
  transaction references, tenant and role checks.

Deploy
- app/sql_migrations.py applies the listed idempotent migrations once each
  during `python -m app.db_bootstrap`; the image now carries them.
- .dockerignore keeps every .env out of the image.

Verified: vitest 173/173, pytest 546/546, audit 138/138, vite build OK.

+++ backend/app/models/appraisals.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import date, datetime
 from typing import Optional
-from sqlalchemy import DateTime, Date, ForeignKey, String, Text, Float, Integer, Boolean
+from sqlalchemy import DateTime, Date, ForeignKey, String, Text, Float, Integer, Boolean, Numeric
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship as orm_relationship
 from sqlalchemy.sql import func
@@ -41,7 +41,7 @@ class StaffAppraisal(Base):
     reviewer_user_id: Mapped[Optional[uuid.UUID]] = mapped_column(UUID(as_uuid=True), nullable=True)
     review_comments: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
     status: Mapped[str] = mapped_column(String, nullable=False, default="pending_review")  # pending_review, approved, rejected
-    salary_increment_pct: Mapped[float] = mapped_column(Float, default=0.0)
+    salary_increment_pct: Mapped[float] = mapped_column(Numeric(8, 3), default=0.0)
     
     created_at: Mapped[Optional[datetime]] = mapped_column(DateTime(timezone=True), server_default=func.now(), nullable=True)
